Phase_1_Advanced/requirements.txt
python-dotenv==1.0.0
boto3==1.28.62
aiohttp==3.8.6
tenacity==8.2.3
beautifulsoup4==4.12.2
aiofiles==23.1.0
